incorporate comments
